floatキャストと動画スキップ

+++ GamePrograming/Application/Source/Game/EnemyCpu.cpp
@@ -111,11 +111,11 @@ void EnemyCpu::Update() {
 		}
 		else
 		{//基本的には通らない//ターゲットPの切り替わりに何フレームか遅延があるためターゲットが死んだ直後だと通るかも
-			stickL.x = m_moveDir;
+			stickL.x = static_cast<float>(m_moveDir);
 			stickL.y = -0.5f;
 		}
 		if (stickL.x == 0 && stickL.y == 0) {//こんな状況はない
-			stickL.x = m_moveDir;
+			stickL.x = static_cast<float>(m_moveDir);
 		}
 	
 	}
@@ -151,7 +151,7 @@ void EnemyCpu::Update() {
 			}
 		}
 
-		stickL.x = m_moveDir;
+		stickL.x = static_cast<float>(m_moveDir);
 	}
 
 	stickL.Normalize();
@@ -347,7 +347,7 @@ void EnemyCpu::Update() {
 		//現在の速度を取得
 		b2Vec2 vel = m_body->GetLinearVelocity();
 		//コントローラーの左右を取得
-		LONG hor = (stickL.x * 1000.0f) * (m_invert ? -1 : 1);
+		LONG hor = static_cast<LONG>((stickL.x * 1000.0f) * (m_invert ? -1 : 1));
 	
 		//コントローラー補正値
 		float controllerCorrection = 0.0f;
